// src/types/visita.ts
export interface Visita {
    id: string;
    idActividad: string;
    inicio: string;
    coordenada: string;
    descripcion: string;
    contacto: string;
    encargado: string;
    foto: string;
    tipo: string;
    pais: string;
    personasAlcanzadas: number;
    contactosNuevos: number;
    kmRecorridos: number;
    tiempoInvertido: number;
    gasolina: number;
    comida: number;
    otrosGastos: number;
    totalInversion: number;
}