// src/types/voluntario.ts
export interface Voluntario {
  id: string;
  fechaInscripcion: string;
  horaInscripcion: string;
  nombreCompleto: string;
  pais: string;
  nacionalidad: string;
  documentoIdentificacion: string;
  fechaNacimiento: string;
  edad: number;
  genero: string;
  profesionOficio: string;
  ocupacion: string;
  telefono: string;
  estadoCivil: string;
  departamentoRegion: string;
  municipioProvincia: string;
  ciudad: string;
  gradoAcademico: string;
  estudiosAdicionales: string;
  correo: string;
  facebook: string;
  instagram: string;
  tiktok: string;
  iglesia: string;
  pastor: string;
  contactoPastor: string;
  denominacion: string;
  areasServicio: string;
  donesTalentos: string;
  cursoStoryline: string;
  diplomaStoryline: string;
  cursoVayaMovilizar: string;
  diplomaVayaMovilizar: string;
  cimasAsistidos: string;
  areasServicioCima: string;
  programasParticipados: string;
  tipoVoluntariado: string;
  testimonio: string;
  cartaPastoral: string;
  foto: string;
  datosCompletos: string;
}